delete-prm.ts had a funky way of determining if a request was in progress or not. Implemented a cleaner way.
Implemented a Quality Service for Requests, which could tell me function based info about a request as it relates to a user. This implementation was interesting because the first implementation, in prm-quality.service, only used promises to return values from its functions. This implementation didn't need to call a service to get its value, the calc was too simple, so it didn't need to return a promise. In fact, returning a promise would have been a bad thing. Implemented code to handle the case of a promise, or an actual value as the return value. Will be refactoring this some more.